requirements.txt
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
Flask-Migrate==2.1.0
Flask-SSLify==0.1.5
Flask-Script==2.0.5
psycopg2==2.7.3
passlib==1.7.1
bcrypt==3.1.4